Initial refactoring of ETL
